% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKSP.R
\name{MKSP}
\alias{MKSP}
\title{Estimate the Multiple Kink effects for Static Panel (MKSP) models.}
\usage{
MKSP(data, iT, var_dep, var_thre, var_base, im, trim_rate = 0.15)
}
\arguments{
\item{data}{data data a tibble of data. The number of rows of \code{data} must be the sample size \code{iT} times individuals number N.}

\item{iT}{the size for time.}

\item{var_dep}{the name of dependent variable.}

\item{var_thre}{the name of threshold variable.}

\item{var_base}{the names for the baseline covariates, i.e. \eqn{x_{it}}. Remark: for static panel model, \code{lag_y} can not be included.}

\item{im}{the number of kink effects.}

\item{trim_rate}{the trimming parameter. Default is 0.15.}
}
\value{
a list holding
\item{beta}{the estimator of the regression coefficients.}
\item{gamma}{the estimator of the threshold parameters.}
\item{beta.se}{the estimated standard error for \code{beta}.}
\item{gamma.se}{the estimated standard error for \code{gamma}}
}
\description{
This function can be used to estimate the multi-kink effects for static panel model using a profile two-stage method develop by Wan et al. (2023). The estimator is also an useful initial value for the minimization of the GMM estimation in MTDP with kink effects.
}
\examples{
N <- 100
iT <- 10
y <- matrix(NA,N,iT)
x1 <- x2 <- matrix(NA,N,iT)
alpha <- rnorm(N,0,1)
rho1 <- runif(N,-0.9,0.9)
rho2 <- runif(N,-0.9,0.9)
for(t in 1:iT){
  xx1 <- if(t>1) x1[,t-1] else 0
  x1[,t] <- rho1*xx1 + 0.25*alpha + rnorm(N,0,0.5)
  xx2 <- if(t>1) x2[,t-1] else 0
  x2[,t] <- rho2*xx2 + 0.5*alpha + rnorm(N,0,0.5)
  yy <- if(t>1) y[,t-1] else 0
  y[,t] <- x1[,t] + x2[,t] + alpha  - 2*(x2[,t] + 0.25)*(x2[,t]> -0.25) + 2*(x2[,t] - 0.25)*(x2[,t]>0.25) + rnorm(N,0,0.5)
}
y <- c(t(y))
x1 <- c(t(x1))
x2 <- c(t(x2))
data = cbind(y,x2,x1)
colnames(data) = c("y","x2","x1")
dep = "y"
thre = "x2"
var_base = c("x1","x2")
MKSP(data,iT,dep,thre,var_base,im=2,trim_rate=0.15)

}
\references{
Wan et al. (2023). Multikink quantile regression for longitudinal data with application to progesterone data analysis. \emph{Biometrics}
Zhang et al. (2017). Panel kink regression with an unknown threshold. \emph{Economics Letter}
Wan and Li. (2025). Dynamic panel models with multi-threshold effects and endogeneity. \emph{Statistics}
}
