% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTDP.R
\name{MTDP}
\alias{MTDP}
\title{: Estimate the Multiple Threshold for Dynamic Panel models.}
\usage{
MTDP(
  use,
  im = 1,
  h_0 = 1,
  par_ini = NULL,
  trim_rate = 0.15,
  avg_num = 0,
  nboots = 0,
  grid_num = 30
)
}
\arguments{
\item{use}{an object of the class MTDP, created by \code{\link{NewMTDP}} function.}

\item{im}{the number of thresholds. The default value is 1.}

\item{h_0}{the constant of the bandwidth for the kernel function, i.e., \eqn{h=h_0\hat{\sigma}n^{-1/5}}. This is mainly used in estimating the variance-covariance matrix of the estimators. Here \code{h_0} can be a vector, and its range is generally [0.1,2). We can set multiple values, and multiple estimated standard deviations will also be output. Default is 1.}

\item{par_ini}{the initial values for the parameters used in the "Nelder-Mead" algorithm. If NULL, the quantiles of threshold variable is specified.}

\item{trim_rate}{the trimming parameter. Default is 0.15.}

\item{avg_num}{the number of randomizations in constructing the weight matrix at the first step of GMM estimation. Default is 0.}

\item{nboots}{the number of bootstrap repetitions in model specification test. If 0,  no test will be performed, and the output p-value will be -1. Default is 0.}

\item{grid_num}{number of grid points in setting the searching domain in the model specification test. It will be used only when \code{nboots} is greater than 0. Default is 30.}
}
\value{
a new object of the class MTDP containing the results from the estimation.

The object is a list containing the components:

\item{J_hat}{the objective value.}
\item{est}{a vector of estimates of all unknown parameters.}
\item{se}{a vector of the standard errors of all the estimates.}
\item{ep_collect}{the estimated residuals.}
\item{L}{the number of instruments}
\item{boots_p}{the p value for the model specification test.}
}
\description{
The function needs the return value (an object of the class MTDP) from the \code{\link{NewMTDP}}. It copies the object, reuses its contents to estimate the corresponding MTDP model, and then returns a new object of the class MTDP containing the results from the estimation. The user can choose to save the return value to a new object or simply to overwrite the object returned from \code{NewMTDP}.
}
\examples{
N <- 100
iT <- 10
y <- matrix(NA,N,iT)
x1 <- x2 <- matrix(NA,N,iT)
alpha <- rnorm(N,0,1)
rho1 <- runif(N,-0.9,0.9)
rho2 <- runif(N,-0.9,0.9)
for(t in 1:iT){
 xx1 <- if(t>1) x1[,t-1] else 0
 x1[,t] <- rho1*xx1 + 0.25*alpha + rnorm(N,0,0.5)
 xx2 <- if(t>1) x2[,t-1] else 0
 x2[,t] <- rho2*xx2 + 0.5*alpha + rnorm(N,0,0.5)
 yy <- if(t>1) y[,t-1] else 0
 y[,t] <- x1[,t] + alpha  + (1-2*x1[,t])*(x2[,t]> -0.25) + (-1+ 2* x1[,t])*(x2[,t]>0.25) + 0.2*yy + rnorm(N,0,0.5)
}
y <- c(t(y))
x1 <- c(t(x1))
x2 <- c(t(x2))
data = cbind(y,x2,x1,1)
colnames(data) = c("y","x2","x1","cons")
dep = "y"
thre = "x2"
var_rx = c("cons","x1")
var_base = c("lag_y","x1")
var_endo = NULL
var_iv = NULL
use = NewMTDP(data,thre_type=c("jump"),dep,thre,var_base,var_rx,var_endo,var_iv,iT)
im = 2
h_0 = c(0.5)
par = NULL
trim_rate=0.15;
grid_num=100;
avg_num=0
method = "grid-search"
obj_mtdp <- MTDP(use,im,h_0,par,trim_rate,avg_num,nboots = 0 )
obj_mtdp$est



}
\references{
Chuang Wan and Yi Li. (2025). Dynamic panel models with multi-threshold effects and endogeneity. \emph{Statistics}
}
\seealso{
\code{\link{NewMTDP}}
}
