% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTSP.R
\name{MTSP}
\alias{MTSP}
\title{Estimate the Multiple Threshold for Static Panel (MTSP) models.}
\usage{
MTSP(
  data,
  iT,
  var_dep,
  var_thre,
  var_base,
  var_rx,
  im = 1,
  trim_rate = 0.15,
  grid_num = 100
)
}
\arguments{
\item{data}{data a tibble of data. The number of rows of \code{data} must be the sample size \code{iT} times individuals number N.}

\item{iT}{the size for time.}

\item{var_dep}{the name of dependent variable.}

\item{var_thre}{the name of threshold variable.}

\item{var_base}{the names for the baseline covariates, i.e. \eqn{x_{it}}. Remark: for static panel model, \code{lag_y} can not be included.}

\item{var_rx}{the names of the regime-dependent variables. For static panel models, \code{lag_y} is not allowed.}

\item{im}{the number of thresholds  (maximum 4). Default value is 1.}

\item{trim_rate}{the trimming parameter. Default is 0.15.}

\item{grid_num}{number of grid points for the grid-search method. Default is 100.}
}
\value{
a list holding
\item{beta}{the estimator of the regression coefficients.}
\item{gamma}{the estimator of the threshold parameters.}
\item{beta.homo.se}{the estimated standard error for \code{beta} with homogeneous error assumption.}
\item{beta.het.se}{the estimated standard error for \code{beta} with heterogeneous error assumption.}
}
\description{
This function mainly implements the sequential estimation procedure proposed by Hansen (1999) designed for static panel model with threshold effects. The estimator obtained by the static panel model is also a useful initial value for the minimization of the GMM estimation in MTDP.
}
\examples{
N <- 100
iT <- 10
y <- matrix(NA,N,iT)
x1 <- x2 <- matrix(NA,N,iT)
alpha <- rnorm(N,0,1)
rho1 <- runif(N,-0.9,0.9)
rho2 <- runif(N,-0.9,0.9)
for(t in 1:iT){
  xx1 <- if(t>1) x1[,t-1] else 0
  x1[,t] <- rho1*xx1 + 0.25*alpha + rnorm(N,0,0.5)
  xx2 <- if(t>1) x2[,t-1] else 0
  x2[,t] <- rho2*xx2 + 0.5*alpha + rnorm(N,0,0.5)
  yy <- if(t>1) y[,t-1] else 0
  y[,t] <- x1[,t] + alpha  + (1-2*x1[,t])*(x2[,t]> -0.25) + (-1+ 2* x1[,t])*(x2[,t]>0.25) + rnorm(N,0,0.5)
}
y <- c(t(y))
x1 <- c(t(x1))
x2 <- c(t(x2))
data = cbind(y,x2,x1,1)
colnames(data) = c("y","x2","x1","cons")
dep = "y"
thre = "x2"
var_rx = c("cons","x1")
var_base = c("x1")
obj_mtsp <- MTSP(data,iT,dep,thre,var_base,var_rx,im=2)
obj_mtsp$beta
obj_mtsp$gamma


}
\references{
Hansen. (1999). Threshold effects in non-dynamic panels: Estimation, testing, and inference. \emph{Journal of Econometrics}
}
