% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewMTDP.R
\name{NewMTDP}
\alias{NewMTDP}
\title{Create an object of the class MTDP (Multiple Threshold for Dynamic Panel models).}
\usage{
NewMTDP(
  data,
  thre_type = c("jump"),
  var_dep,
  var_thre,
  var_base,
  var_rx,
  var_endo,
  var_iv = NULL,
  iT
)
}
\arguments{
\item{data}{a tibble of data. The number of rows of \code{data} must be the sample size \code{iT} times individuals number N.}

\item{thre_type}{the threshold regression type. Options are
\describe{
\item{\code{"jump"}}{for estimating the threshold regression with jump effects.}
\item{\code{"kink"}}{for estimating the kink regression with continuous threshold effects.}
} Default is \code{"jump"}.}

\item{var_dep}{the name of dependent variable.}

\item{var_thre}{the name of threshold variable.}

\item{var_base}{the names for the baseline covariates, i.e. \eqn{x_{it}}. Remark: for dynamic model, \code{lag_y} should be included.}

\item{var_rx}{the names of the regime-dependent variables; the \code{lag_y} and the \code{intercept} are allowed, but need to be specified.}

\item{var_endo}{the names for all endogenous variables including the regressors and threshold variables.}

\item{var_iv}{the names for additional instrument variables. Default is NULL.}

\item{iT}{the size for time.}
}
\value{
An object of the class MTDP for later usage in \code{\link{MTDP}}.
}
\description{
Create an object of the class MTDP (Multiple Threshold for Dynamic Panel models).
}
\examples{

N <- 100
iT <- 10
y <- matrix(NA,N,iT)
x1 <- x2 <- matrix(NA,N,iT)
alpha <- rnorm(N,0,1)
rho1 <- runif(N,-0.9,0.9)
rho2 <- runif(N,-0.9,0.9)
for(t in 1:iT){
  xx1 <- if(t>1) x1[,t-1] else 0
 x1[,t] <- rho1*xx1 + 0.25*alpha + rnorm(N,0,0.5)
 xx2 <- if(t>1) x2[,t-1] else 0
 x2[,t] <- rho2*xx2 + 0.5*alpha + rnorm(N,0,0.5)
 yy <- if(t>1) y[,t-1] else 0
 y[,t] <- x1[,t] + alpha  + (1-2*x1[,t])*(x2[,t]> -0.25) + (-1+ 2* x1[,t])*(x2[,t]>0.25) + 0.2*yy + rnorm(N,0,0.5)
}
y <- c(t(y))
x1 <- c(t(x1))
x2 <- c(t(x2))
data = cbind(y,x2,x1,1)
colnames(data) = c("y","x2","x1","cons")
dep = "y"
thre = "x2"
var_rx = c("cons","x1")
var_base = c("lag_y","x1")
var_endo = NULL
var_iv = NULL
use = NewMTDP(data,dep,thre,var_base,var_rx, var_endo,var_iv,iT)


}
\references{
Wan and Li. (2025). Dynamic panel models with multi-threshold effects and endogeneity. \emph{Statistics}
}
\seealso{
\code{\link{MTDP}}
}
